@@ -2,6 +2,7 @@
     "name": "context-please",
     "version": "0.1.0",
     "description": "A powerful code indexing tool with multi-platform support (fork of claude-context)",
+    "type": "module",
     "private": true,
     "scripts": {
         "build": "pnpm -r --filter='!./examples/*' build && pnpm -r --filter='./examples/*' build",
@@ -25,12 +26,14 @@
         "benchmark": "node scripts/build-benchmark.js"
     },
     "devDependencies": {
+        "@antfu/eslint-config": "^5.4.1",
         "@types/node": "^20.0.0",
         "@typescript-eslint/eslint-plugin": "^8.31.1",
         "@typescript-eslint/parser": "^8.31.1",
         "assert": "^2.1.0",
         "browserify-zlib": "^0.2.0",
-        "eslint": "^9.25.1",
+        "eslint": "^9.37.0",
+        "eslint-plugin-format": "^1.0.2",
         "https-browserify": "^1.0.0",
         "os-browserify": "^0.3.0",
         "rimraf": "^6.0.1",
